Complete refactoring Phases 2 & 3: split checker.py + extract service layer
Phase 2: Split checker.py (3,939 → 340 lines) into focused modules:
- diff_types.py (915 lines): type/struct/enum/union/typedef detectors
- diff_platform.py (1,205 lines): ELF/PE/Mach-O/DWARF platform detectors
- diff_filtering.py (1,002 lines): enrichment, redundancy filtering, dedup
- diff_symbols.py: added 4 missing param/var detectors from checker.py
- checker.py: now only orchestration (compare function + detector registry)

Phase 3: Extract service.py as shared orchestration layer:
- resolve_input(): auto-detect and load ABI snapshots from any format
- run_dump()/run_compare(): framework-agnostic binary dump and comparison
- render_output(): format-agnostic output rendering (JSON/MD/SARIF/HTML)
- Uses project error hierarchy (SnapshotError, ValidationError) instead
  of framework-specific exceptions (click.ClickException)

All 3,464 tests pass. Zero test file changes. mypy clean.
